+++ test/application_system_test_case.rb
@@ -4,4 +4,12 @@ require 'test_helper'
 
 class ApplicationSystemTestCase < ActionDispatch::SystemTestCase
   driven_by :selenium, using: :chrome, screen_size: [1400, 1400]
+
+  def login_user(user)
+    visit welcome_url
+    OmniAuth.config.mock_auth[:discord] = nil
+    Rails.application.env_config['omniauth.auth'] = discord_mock(user.name, user.uid)
+    stub_request(:get, "#{Discordrb::API.api_base}/guilds/#{ENV['DISCORD_SERVER_ID']}/members/#{user.uid}")
+    click_link 'ログイン'
+  end
 end
