--- test/models/question_test.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require 'test_helper'
-
-class QuestionTest < ActiveSupport::TestCase
-  # test "the truth" do
-  #   assert true
-  # end
-end
